Update Dockerfile and docker-compose.yml configurations
The docker configurations have been modified for better package handling. In the Dockerfile, the environment has been updated to python:3.11-alpine3.20, the method of package installation is also adjusted. Moreover, in docker-compose.yml, command execution formatting is refined and volume mapping is included for sessions and environment file is added for better configuration setup.