Update Docker configurations
Updated Dockerfile and docker-compose.yml file. The Dockerfile is now using a builder pattern for image creation, and the python version has been updated. The bot service in the docker-compose.yml now receives the command as a list, the volumes have been adjusted, and it reads environment variables from a .env file in the same directory.