Update Docker setup and improve testing workflow
This update changes the main Docker setup to use a pre-built image instead of building every time the container starts. It also introduces an environment variable with a default value. The Github Actions workflow has been adjusted to match these changes and now includes better container testing steps. This enhances the reliability of the container and makes testing more efficient.