Update Docker build-push workflow
The Github Actions workflow for building and pushing Docker images has been updated. This includes the removal of the step for repository checkout and modifications to the metadata labels for the Docker images. The workflow now uses Docker build-push-action@v6 and the Github token for authentication instead of a custom token.